requirements.txt
anyio==4.3.0
cachetools==5.3.3
certifi==2024.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-dotenv==1.0.1
python-telegram-bot==21.1.1
sniffio==1.3.1
TgCrypto==1.2.5
